Add txEncoder arg to Check & Deliver
