+++ x/bank/bench_test.go
@@ -46,12 +46,12 @@ func BenchmarkOneBankSendTxPerBlock(b *testing.B) {
 	// Committing, and what time comes from Check/Deliver Tx.
 	for i := 0; i < b.N; i++ {
 		benchmarkApp.BeginBlock(abci.RequestBeginBlock{Header: tmproto.Header{Height: height}})
-		_, _, err := benchmarkApp.Check(txs[i])
+		_, _, err := benchmarkApp.Check(txGen.TxEncoder(), txs[i])
 		if err != nil {
 			panic("something is broken in checking transaction")
 		}
 
-		_, _, err = benchmarkApp.Deliver(txs[i])
+		_, _, err = benchmarkApp.Deliver(txGen.TxEncoder(), txs[i])
 		require.NoError(b, err)
 		benchmarkApp.EndBlock(abci.RequestEndBlock{Height: height})
 		benchmarkApp.Commit()
@@ -88,12 +88,12 @@ func BenchmarkOneBankMultiSendTxPerBlock(b *testing.B) {
 	// Committing, and what time comes from Check/Deliver Tx.
 	for i := 0; i < b.N; i++ {
 		benchmarkApp.BeginBlock(abci.RequestBeginBlock{Header: tmproto.Header{Height: height}})
-		_, _, err := benchmarkApp.Check(txs[i])
+		_, _, err := benchmarkApp.Check(txGen.TxEncoder(), txs[i])
 		if err != nil {
 			panic("something is broken in checking transaction")
 		}
 
-		_, _, err = benchmarkApp.Deliver(txs[i])
+		_, _, err = benchmarkApp.Deliver(txGen.TxEncoder(), txs[i])
 		require.NoError(b, err)
 		benchmarkApp.EndBlock(abci.RequestEndBlock{Height: height})
 		benchmarkApp.Commit()
